+++ CHANGELOG.md
@@ -30,10 +30,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Upcoming release] - 1971-11-03
 
 ### Added
--
+- Add external link example to the subject listing section
 
 ### Changed
--
+- Remove underline on external links in subject listing
 
 ### Deprecated
 -